+++ react/src/index.js
@@ -19,4 +19,5 @@ export { Option, OptionsMenu } from './components/Options';
 export { Table, TableWrap, TableCell, TableCellAction, KebabIcon } from './components/Table';
 export { Snackbar, SnackbarClose } from './components/Snackbar';
 export { ProgressBar } from './components/ProgressBar';
+export { Link } from './components/Link';
 export { default } from './components/Button';